added params for tableView
widget.newTableView added, also when using newTableView and when calling
 sync(tableData). I made a table for populating tableViews, simply type
tableData = { } or local tableData = { } and you will get all the curren
t params available inside the { }.

Latest updates to tableView, a few n
ew features were added. onLeftSwipe, onRightSwipe, deleteRow() and scrol
lToTop(). Check the documentation on that in the API section on the Ansc
a website.